+++ utils/common_vars.py
@@ -11,3 +11,4 @@ DATA_DIRECTORY = f'{USER_DIRECTORY}/generate_data/'
 FLIGHTS = 'flights'
 BUS = 'bus'
 TRAIN = 'train'
+transportation_type_list = [FLIGHTS, BUS, TRAIN]
