+++ utils/get_csv_data.py
@@ -4,11 +4,9 @@ import argparse
 import os
 import sys
 import pandas as pd
-from common_vars import TRAIN, BUS, FLIGHTS, DATA_DIRECTORY
+from common_vars import DATA_DIRECTORY
 from common_funcs import get_verbose, get_transportation_type, get_on_aws, get_aws_profile, get_bucket, \
-    get_s3_client, check_if_object_exists_in_s3, read_object_from_s3
-
-transportation_type_list = [FLIGHTS, BUS, TRAIN]
+    get_s3_client, check_if_object_exists_in_s3, read_object_from_s3, transport_in_list
 
 
 def check_local_exist(transportation_type):
@@ -17,37 +15,38 @@ def check_local_exist(transportation_type):
     return False
 
 
+def handle_df(obj):
+    rows = obj.split('\n')
+    df = pd.DataFrame(rows)
+    df = df[0].str.split(',', expand=True)
+    df.columns = df.iloc[0]
+    df.drop(df.index[0], inplace=True)
+    return df
+
+
 def get_csv_data(transportation_type, aws_profile, on_aws, bucket):
+    df = pd.DataFrame()
     try:
-        if transportation_type not in transportation_type_list:
-            verboseprint(
-                f'Error in {transportation_type} - Invalid transportation type')
-            return False, pd.DataFrame()
-        df = pd.DataFrame()
+        if not transport_in_list:
+            return False, df
         if on_aws:
             s3_client = get_s3_client(aws_profile)
-            if check_if_object_exists_in_s3(bucket, f'{transportation_type}.csv', s3_client):
+            if check_if_object_exists_in_s3(bucket, f'{transportation_type}.csv', s3_client=get_s3_client):
                 verboseprint(
                     f'Object {transportation_type}.csv exists in S3, retrieving from S3...')
                 obj = read_object_from_s3(
                     bucket, f'{transportation_type}.csv', s3_client)
-                rows = obj.split('\n')
-                df = pd.DataFrame(rows)
-                df = df[0].str.split(',', expand=True)
-                df.columns = df.iloc[0]
-                df.drop(df.index[0], inplace=True)
+                df = handle_df(obj)
                 verboseprint(df)
                 return True, df
-            return False, df
-        else:
-            if check_local_exist(transportation_type):
-                verboseprint(
-                    f'Object {transportation_type}.csv exists locally, retrieving from local...')
-                df = pd.read_csv(
-                    f'{DATA_DIRECTORY}{transportation_type}.csv', encoding='utf-8')
-                verboseprint(df)
-                return True, df
-            return False, df
+        elif check_local_exist(transportation_type):
+            verboseprint(
+                f'Object {transportation_type}.csv exists locally, retrieving from local...')
+            df = pd.read_csv(
+                f'{DATA_DIRECTORY}{transportation_type}.csv', encoding='utf-8')
+            verboseprint(df)
+            return True, df
+        return False, df
     except Exception as e:
         verboseprint(f'Error in get_csv_data() - {e}')
         return False, df
@@ -73,7 +72,6 @@ def main():
             verboseprint(f'{transportation_type}.csv has successful retrieved')
         else:
             verboseprint(f'{transportation_type}.csv has failed to retrieve')
-    return
 
 
 def check_args(args=None):
